requirements.txt
requests_oauthlib
load_dotenv
